feat: sevidor_temporario : Criado Entity e repository

+++ src/main/resources/db/migration/V0010__criar_tabela_servidor_temporario.sql
@@ -1,5 +1,5 @@
 CREATE TABLE servidor_temporario (
-    pes_id INT NOT NULL,
+    pes_id bigint NOT NULL,
     st_data_admissao DATE NOT NULL,
     st_data_demissao DATE NULL,
     PRIMARY KEY (pes_id, st_data_admissao),
